Add files via upload
This notebook uses the Kaggle dataset of Lego Minifigurine images. The aim is to train a model to recognize the figures (which Star Wars character they depict).
I have first tried a convolutional neural net model from scratch (called baby_lego.h5). Baby_lego had an accuracy of 0.9 for training but the validation accuracy was 0.6. But it performad well with the test dataset (0.9 as well).
Then I have tried a transfer learning strategy with the VGG model (only modifying and training only the last layer). I called this model padawan.h5. padawan had an accuracy of 1 for the test dataset.